requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==23.0.0
packaging==25.0
Pillow==10.1.0
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.14.1
whitenoise==6.9.0
